refactor remoteSSH to not subclass Filesystem
